Remove separate `expiry` field from stored OAuth credential and instead parse it out of the PASETO token as needed
